+++ wcfsetup/install/files/lib/data/user/UserAction.class.php
@@ -328,7 +328,7 @@ class UserAction extends AbstractDatabaseObjectAction implements IClipboardActio
 				}
 			}
 			
-			if (array_key_exists('password', $this->parameters['data'])) {
+			if (array_key_exists('password', $this->parameters['data']) && $this->parameters['data']['password'] !== '') {
 				foreach ($this->getObjects() as $object) {
 					SessionHandler::getInstance()->deleteUserSessionsExcept($object->getDecoratedObject(), SessionHandler::getInstance()->sessionID);
 					SessionHandler::getInstance()->deleteAcpSessionsExcept($object->getDecoratedObject(), SessionHandler::getInstance()->sessionID);
